Attempt to handle CFF fonts with other than 1000 units per em.
Marc Penninga reported one.

+++ otftotfm/otftotfm.hh
@@ -17,9 +17,9 @@ void output_metrics(Metrics &metrics, const String &ps_name, int boundary_char,
 	String (*dvips_include)(const String &ps_name, const FontInfo &, ErrorHandler *),
 	ErrorHandler *errh);
 
-int font_x_height(const FontInfo &, const Transform &);
-int font_cap_height(const FontInfo &, const Transform &);
-int font_ascender(const FontInfo &, const Transform &);
+double font_x_height(const FontInfo &, const Transform &);
+double font_cap_height(const FontInfo &, const Transform &);
+double font_ascender(const FontInfo &, const Transform &);
 double font_slant(const FontInfo &);
 
 #endif
